sourcemeter-ui/requirements.txt
fastapi==0.88.0
starlette==0.22.0
sqlalchemy==2.0.23
uvicorn[standard]
psycopg2
kafka-python
docker
jinja2
